a few fixes... - reduced unnecessary copying of Chrome's history file   (it was copying hundreds of megabytes every 10 seconds) - started adding python3 support - made script keep running even if Chrome history access fails - ignore Discord channel URLs
Tried to eliminate the need to copy the history file, but... it seems completely infeasible.
The internet in general hasn't managed to find a solution for it.
The best method appears to be using a completely different approach, like
making a browser extension which sends URLs real-time to a web service.  Basically,
do something similar to the existing queue system, but via a web service instead of
a local file.  And make an extension per browser.  Might be a good plan for the future.